+++ tests/property_tests/_types.py
@@ -18,6 +18,7 @@ MyNamedOrderedBidict = namedbidict('MyNamedOrderedBidict', 'key', 'val', base_ty
 MUTABLE_BIDICT_TYPES = (bidict, OrderedBidict, MyNamedBidict)
 FROZEN_BIDICT_TYPES = (frozenbidict, FrozenOrderedBidict, MyNamedFrozenBidict)
 ORDERED_BIDICT_TYPES = (OrderedBidict, FrozenOrderedBidict, MyNamedOrderedBidict)
+ORDER_PRESERVING_BIDICT_TYPES = tuple(set(FROZEN_BIDICT_TYPES + ORDERED_BIDICT_TYPES))
 BIDICT_TYPES = tuple(set(MUTABLE_BIDICT_TYPES + FROZEN_BIDICT_TYPES + ORDERED_BIDICT_TYPES))
 # When support is dropped for Python < 3.8, all bidict types will be reversible,
 # and we can remove the following and just use BIDICT_TYPES instead: